+++ src/atc/configurator/sql/exceptions.py
@@ -1,0 +1,5 @@
+from atc.exceptions.configurator_exceptions import AtcConfiguratorInvalidSqlException
+
+
+class _UnpackAttemptFailed(AtcConfiguratorInvalidSqlException):
+    pass
